+++ Room.java
@@ -1,14 +1,14 @@
 package edu.baylor.hoteltransylvania;
 
 
-import edu.baylor.hoteltransylvania.BedType;
-import edu.baylor.hoteltransylvania.QualityLevel;
-import edu.baylor.hoteltransylvania.RoomStatus;
-import edu.baylor.hoteltransylvania.RoomType;
+//import RoomEnums.BedType;
+//import RoomEnums.QualityLevel;
+//import RoomEnums.RoomStatus;
+//import RoomEnums.RoomType;
 
 //import java.util.Date;
 
-class Room {
+public class Room {
     // Constants (inclusive values)
     public static final Integer
             FIRST_FLOOR_MIN = 100, FIRST_FLOOR_MAX = 135,
@@ -141,8 +141,8 @@ class Room {
     }
     
     public boolean match(Room rm) {
-    	if (roomStatus.equals(rm.roomStatus) && bedType.equals(rm.bedType) 
-    	 && roomType.equals(rm.roomType) && qualityLevel.equals(rm.qualityLevel) && smokingAllowed.equals(rm.smokingAllowed)) {
+    	if (bedType.equals(rm.bedType) && roomType.equals(rm.roomType) && 
+    		qualityLevel.equals(rm.qualityLevel) && smokingAllowed.equals(rm.smokingAllowed)) {
     		return true;
     	}
     	return false;